fix: address code review findings for aura-gemini
- Reset reconnection attempt counter after successful session
- Add ToolCallCancellation event variant and handler
- Make goAway trigger reconnection instead of being silently logged
- Add base64 decode error logging for corrupted audio frames
- Collapse nested if into let-chain (clippy)
- Add explicit summarize_screen => None match arm in tools.rs
- Align temperature type to f64 in config.rs

+++ crates/aura-gemini/src/config.rs
@@ -12,7 +12,7 @@ pub struct GeminiConfig {
     pub model: String,
     pub voice: String,
     pub system_prompt: String,
-    pub temperature: f32,
+    pub temperature: f64,
     pub ws_url_override: Option<String>,
 }
